Lab 8 done

+++ app/views/calculator.py
@@ -5,6 +5,13 @@ from flask import Blueprint, jsonify, request
 import app.models as models
 import app.db as db
 from flask_bcrypt import Bcrypt
+from app.views.user import UserRole
+from app import app
+from flask_jwt_extended import (
+    JWTManager, jwt_required, get_jwt_identity
+)
+
+jwt = JWTManager(app)
 
 calculator_blpr = Blueprint('calculator', __name__, url_prefix='/calculator')
 calculators_blpr = Blueprint('calculators', __name__, url_prefix='/calculators')
@@ -31,6 +38,7 @@ def get_all_calculators():
 	return jsonify(res), 200
 
 @calculator_blpr.route('/', methods=['POST'])
+@jwt_required()
 def create_new_calculator():
 	class Calculator(Schema):
 		name = fields.Str(required=True)
@@ -38,21 +46,20 @@ def create_new_calculator():
 		inputData = fields.Str(required=True)
 		code = fields.Str(required=True)
 		isPublic = fields.Bool(required=True)
-		authorId = fields.Int(required=True)
 	try:
 		if not request.json:
 			raise ValidationError('No input data provided')
-		Calculator().validate(request.json)
+		calculator = Calculator().load(request.json)
 	except ValidationError as err:
 		return jsonify(err.messages), 400
 
 	calculator = models.Calculator(
-		name=request.json['name'],
-		description=request.json['description'],
-		input_data=request.json['inputData'],
-		code=request.json['code'],
-		is_public=request.json['isPublic'],
-		author_id=request.json['authorId'])
+		name=calculator['name'],
+		description=calculator['description'],
+		input_data=calculator['inputData'],
+		code=calculator['code'],
+		is_public=calculator['isPublic'],
+		author_id=get_jwt_identity())
 	
 	res = {}
 	try:
@@ -112,6 +119,7 @@ def get_calculator(calculator_id):
 	return jsonify(res), 200
 
 @calculator_blpr.route('/<int:calculator_id>', methods=['PATCH'])
+@jwt_required()
 def update_calculator(calculator_id):
 
 	class Calculator(Schema):
@@ -124,7 +132,7 @@ def update_calculator(calculator_id):
 	try:
 		if not request.json:
 			raise ValidationError('No input data provided')
-		Calculator().validate(request.json)
+		calculator = Calculator().load(request.json)
 	except ValidationError as err:
 		return jsonify(err.messages), 400
 		
@@ -132,17 +140,22 @@ def update_calculator(calculator_id):
 	if calculator_model is None:
 		return jsonify({'error': 'Calculator not found'}), 404
 
+	user = db.session.query(models.User).filter_by(id=get_jwt_identity()).first()
+	if calculator_model.author_id != get_jwt_identity() and user.role != UserRole.Administrator and user.role != UserRole.Moderator:
+		return jsonify({'error': 'Access denied'}), 403
+
+
 	try:
-		if 'name' in request.json:
-			calculator_model.name = request.json['name']
-		if 'description' in request.json:
-			calculator_model.description = request.json['description']
-		if 'inputData' in request.json:
-			calculator_model.input_data = request.json['inputData']
-		if 'code' in request.json:
-			calculator_model.code = request.json['code']
-		if 'isPublic' in request.json:
-			calculator_model.is_public = request.json['isPublic']
+		if 'name' in calculator:
+			calculator_model.name = calculator['name']
+		if 'description' in calculator:
+			calculator_model.description = calculator['description']
+		if 'inputData' in calculator:
+			calculator_model.input_data = calculator['inputData']
+		if 'code' in calculator:
+			calculator_model.code = calculator['code']
+		if 'isPublic' in calculator:
+			calculator_model.is_public = calculator['isPublic']
 	except:
 		db.session.rollback()
 		return jsonify({'error': 'Database error'}), 500
@@ -152,10 +165,15 @@ def update_calculator(calculator_id):
 	return get_calculator(calculator_id)
 
 @calculator_blpr.route('/<int:calculator_id>', methods=['DELETE'])
+@jwt_required()
 def delete_calculator(calculator_id):
 	calculator_model = db.session.query(models.Calculator).filter_by(id=calculator_id).first()
 	if calculator_model is None:
 		return jsonify({'error': 'Calculator not found'}), 404
+
+	user = db.session.query(models.User).filter_by(id=get_jwt_identity()).first()
+	if calculator_model.author_id != get_jwt_identity() and user.role != UserRole.Administrator:
+		return jsonify({'error': 'Access denied'}), 403
 
 	try:
 		db.session.delete(calculator_model)
@@ -187,6 +205,7 @@ def get_all_reviws_for_calculator(calculator_id):
 	return jsonify(res), 200
 
 @calculators_blpr.route('/<int:review_calculator_id>/reviews', methods=['POST'])
+@jwt_required()
 def add_review_to_calculator(review_calculator_id):
 	if db.session.query(models.Calculator).filter_by(id=review_calculator_id).first() is None:
 		return jsonify({'error': 'Calculator not found'}), 404
@@ -199,11 +218,11 @@ def add_review_to_calculator(review_calculator_id):
 	try:
 		if not request.json:
 			raise ValidationError('No input data provided')
-		Review().validate(request.json)
+		review = Review().load(request.json)
 	except ValidationError as err:
 		return jsonify(err.messages), 400
 
-	if not(0 <= request.json['rating'] <= 5):
+	if not(0 <= review['rating'] <= 5):
 		return jsonify({'error': 'Rating must be between 0 and 5'}), 400
 
 	old_review = db.session.query(
